requirements.txt
pandas 
llama-index
pymilvus
openai==0.28
python-dotenv